+++ src/utils/date.js
@@ -1,0 +1,10 @@
+import {MONTHS} from "../const";
+
+export const formatDate = (date) => `${date.getDate()} ${MONTHS[date.getMonth()]}`;
+
+export const formatTime = (date) => {
+  const hours = String(date.getHours()).padStart(2, `0`);
+  const minutes = String(date.getMinutes()).padStart(2, `0`);
+
+  return `${hours}:${minutes}`;
+};
